move test_convert_points_to_homogeneous to conversions tests
