+++ .gitignore
@@ -7,6 +7,7 @@ torchgeometry.egg-info/
 *~
 *.swp
 *.swo
+*.swn
 .eggs/*
 .cache/*
 .pytest_cache/*
